LayoutTests/fast/table/table-with-borderattr-set-to-null.html

<!DOCTYPE html>
<head>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();
</script>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
    <p>There should be a black box below. <br> 

    <table id="table" width="100%" height="100%">
        <tr>
            <td height="26"></td>
        </tr>
    </table>

    <div id="console"></div>
    <script>
        var table = document.getElementById("table");
        table.setAttribute("border", null);
        shouldBeEqualToString('getComputedStyle(document.querySelector("table")).borderTopWidth', "1px");
    </script>
</body>
